Class-based view for Django
The view `ElasticListView` behaves exactly like Django's original
`ListView`, but operates on an ElasticSearch `Search` instance instead
of a Django `QuerySet`.

This implies that the `Search` instance needs to be paginated before
converting to a Django `QuerySet` that is used for template rendering.

Exceptions related to connection errors with the ElasticSearch API, are
handled and corresponding error messages are passed to the client by
Django's messages middleware.