Add support for --refresh to search_index command
Currently when updating or rebuilding an index, the `refresh` parameter
is never set, regardless if settings.ELASTICSEARCH_DSL_AUTO_REFRESH is
set or if auto_refresh is set on the Document class.

This commit adds the ability to pass `--refresh` to the `search_index`
management command, to force a refresh when the index is updated.

This can be useful, for example, if you want to programmatically call
that command from some Django tests that fill an index, and then
immediately run tests against that index. Without setting
`refresh=True`, there's no guarantee that the refreshed index will be
available to search.